@@ -24,7 +24,7 @@
     "url": "https://github.com/CoxAutomotiveDataSolutions/react-typewriter.git"
   },
   "peerDependencies": {
-    "react": "^0.14.7"
+    "react": "^16.2.0"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
@@ -41,5 +41,8 @@
     "react-hot-loader": "^1.2.7",
     "webpack": "^2.1.0-beta.22",
     "webpack-dev-server": "^1.10.1"
+  },
+  "dependencies": {
+    "prop-types": "^15.6.2"
   }
 }
